feat: split the component into a SearchFilter
